backend/ml/requirements-train.txt
ultralytics==8.3.70
opencv-python==4.10.0.84
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
pyyaml==6.0.3
tqdm==4.67.1
